update directory and file structure
