+++ lib/qonversion_flutter.dart
@@ -1,0 +1,2 @@
+export 'src/qonversion.dart';
+export 'src/qa_provider.dart';
